gemini/requirements.txt
google-generativeai==0.3.2
pandas==2.1.4
numpy==1.24.3
typer==0.9.0
result==0.16.0
websockets==12.0
tensorflow-rocm==2.13.0.570
